+++ test/CosmosDB.permissions.Tests.ps1
@@ -68,7 +68,7 @@ InModuleScope CosmosDB {
             { Get-Command -Name Get-CosmosDbPermission } | Should -Not -Throw
         }
 
-        Context 'Called with connection parameter and an id' {
+        Context 'Called with connection parameter and no id' {
             Mock `
                 -CommandName Invoke-CosmosDbRequest `
                 -ParameterFilter { $Method -eq 'Get' -and $ResourceType -eq 'permissions' } `
@@ -95,7 +95,7 @@ InModuleScope CosmosDB {
             }
         }
 
-        Context 'Called with connection parameter and no id' {
+        Context 'Called with connection parameter and an id' {
             Mock `
                 -CommandName Invoke-CosmosDbRequest `
                 -ParameterFilter { $Method -eq 'Get' -and $ResourceType -eq 'permissions' -and $ResourcePath -eq ('users/{0}/permissions/{1}' -f $script:testUser, $script:testPermission) } `